+++ src/core/integer.ts
@@ -1,4 +1,5 @@
-import { MaxSizeError, MinSizeError, NotIntegerError, RawTypeError, LogicError } from './errors'
+import { isDefined } from '../utils'
+import { LogicError, MaxSizeError, MinSizeError, NotIntegerError, RawTypeError } from './errors'
 
 export interface VOIntegerOptions {
   /**
@@ -69,13 +70,13 @@ export interface VOIntegerConstructor {
  * ```
  */
 export const VOInteger = (options: VOIntegerOptions = {}): VOIntegerConstructor => {
-  if (options.min !== undefined) {
+  if (isDefined(options.min)) {
     if (typeof options.min !== 'number') throw new RawTypeError('number', typeof options.min, 'options.min')
   }
-  if (options.max !== undefined) {
+  if (isDefined(options.max)) {
     if (typeof options.max !== 'number') throw new RawTypeError('number', typeof options.max, 'options.max')
   }
-  if (options.min !== undefined && options.max !== undefined) {
+  if (isDefined(options.min) && isDefined(options.max)) {
     if (options.min > options.max) throw new LogicError('options.min should not be bigger than options.max')
   }
 
@@ -84,8 +85,8 @@ export const VOInteger = (options: VOIntegerOptions = {}): VOIntegerConstructor 
 
     constructor(raw: number) {
       if (typeof raw !== 'number') throw new RawTypeError('number', typeof raw, 'raw')
-      if (options.min !== undefined && raw < options.min) throw new MinSizeError(options.min, raw)
-      if (options.max !== undefined && raw > options.max) throw new MaxSizeError(options.max, raw)
+      if (isDefined(options.min) && raw < options.min) throw new MinSizeError(options.min, raw)
+      if (isDefined(options.max) && raw > options.max) throw new MaxSizeError(options.max, raw)
       if (!Number.isInteger(raw)) throw new NotIntegerError(raw)
       this._value = raw
     }