+++ src/e2e/object.spec.ts
@@ -1,9 +1,12 @@
-import { VOObject, VOObjectOptions } from './object'
+import { expectTypeOf } from 'expect-type'
+import { VOError, VOObject, VOObjectOptions } from '../..'
+import { isNull } from '../utils'
+import { constructorFn } from './utils'
 
 describe('VOObject', () => {
   class AAA {
     constructor(raw: 123) {
-      if (raw === null) throw Error('AAA test error')
+      if (isNull(raw)) throw new VOError('AAA test error')
     }
     valueOf(): 'aaa' {
       return 'aaa'
@@ -12,7 +15,7 @@ describe('VOObject', () => {
 
   class BBB {
     constructor(raw: 456) {
-      if (raw === null) throw Error('BBB test error')
+      if (isNull(raw)) throw new VOError('BBB test error')
     }
     valueOf(): 'bbb' {
       return 'bbb'
@@ -21,7 +24,7 @@ describe('VOObject', () => {
 
   class CCC {
     constructor(raw: 789) {
-      if (raw === null) throw Error('CCC test error')
+      if (isNull(raw)) throw new VOError('CCC test error')
     }
     valueOf(): 'ccc' {
       return 'ccc'
@@ -81,9 +84,8 @@ describe('VOObject', () => {
         (errArray): boolean =>
           Array.isArray(errArray) &&
           errArray.length === test.props.length &&
-          errArray.every(err => err instanceof Error) &&
-          errArray.every(err => typeof err.prop === 'string') &&
-          errArray.every(err => test.props.includes(err.prop)),
+          errArray.every(VOError.is) &&
+          errArray.every(err => test.props.includes(<string>err.path.toArray().pop())),
       )
     }
   })
@@ -100,7 +102,7 @@ describe('VOObject', () => {
 
     for (const test of tests) {
       const fn = () => VOObject({ aaa: AAA }, test)
-      if (test.error === null) expect(fn).not.toThrow()
+      if (isNull(test.error)) expect(fn).not.toThrow()
       else expect(fn).toThrowError(test.error)
     }
   })
@@ -110,7 +112,7 @@ describe('VOObject', () => {
 
     class Base {
       constructor(raw: number) {
-        throw Error('Test error')
+        throw new VOError('Test error')
       }
       valueOf(): 'base test' {
         return 'base test'
@@ -140,11 +142,22 @@ describe('VOObject', () => {
         return (
           Array.isArray(errArray) &&
           errArray.length === maxErrors &&
-          errArray.every(err => err instanceof Error) &&
-          errArray.every(err => typeof err.prop === 'string') &&
-          errArray.every(err => props.has(err.prop))
+          errArray.every(VOError.is) &&
+          errArray.every(err => props.has(<string>err.path.toArray().pop()))
         )
       })
     }
+  })
+
+  it('Should have the correct types', () => {
+    class Test extends VOObject({ aaa: AAA, bbb: BBB, ccc: CCC }) {}
+    const instance = new Test({ aaa: 123, bbb: 456, ccc: 789 })
+
+    expectTypeOf(constructorFn(Test)).toEqualTypeOf<(r: { aaa: 123; bbb: 456; ccc: 789 }) => Test>()
+    expectTypeOf(instance.aaa).toEqualTypeOf<AAA>()
+    expectTypeOf(instance.bbb).toEqualTypeOf<BBB>()
+    expectTypeOf(instance.ccc).toEqualTypeOf<CCC>()
+    expectTypeOf(instance.valueOf()).toEqualTypeOf<{ aaa: 'aaa'; bbb: 'bbb'; ccc: 'ccc' }>()
+    expectTypeOf(instance.toRaw()).toEqualTypeOf<{ aaa: 'aaa'; bbb: 'bbb'; ccc: 'ccc' }>()
   })
 })