+++ src/e2e/optional.spec.ts
@@ -1,5 +1,5 @@
-import { VOOptional } from './optional'
-import { RawTypeError } from './errors'
+import { expectTypeOf } from 'expect-type'
+import { Noneable, RawTypeError, VOOptional } from '../..'
 
 describe('VOOptional', () => {
   class Base {
@@ -88,7 +88,24 @@ describe('VOOptional', () => {
       expect(instance.value).toBe(test.raw)
     }
   })
+
+  it('Should have the correct types', () => {
+    class Test {
+      constructor(raw: number) {}
+      valueOf(): string {
+        return ''
+      }
+    }
+
+    // !WARNING: Removing `undefined` from some expected types due to https://github.com/mmkal/ts/issues/187
+    // TODO: Enable `undefined` once https://github.com/mmkal/ts/issues/187 is fixed
+    expectTypeOf(new (VOOptional(Test, [null, undefined]))(434).valueOf()).toEqualTypeOf<string | null>()
+    expectTypeOf(new (VOOptional(Test, [null]))(123).valueOf()).toEqualTypeOf<string | null>()
+    expectTypeOf(new (VOOptional(Test, [undefined]))(434).valueOf()).toEqualTypeOf<string>()
+    expectTypeOf(new (VOOptional(Test, []))(434).valueOf()).toEqualTypeOf<string>()
+    expectTypeOf(new (VOOptional(Test))(434).valueOf()).toEqualTypeOf<string>()
+  })
 })
 
-const noneableValues = [null, undefined]
+const noneableValues: Array<Noneable> = [null, undefined]
 const nonNoneableValues = [0, false, []]