+++ src/lib/email.ts
@@ -1,4 +1,0 @@
-import { VOString } from '../core'
-
-const EMAIL_PATTERN = /^(?=.{1,254}$)(?=.{1,64}@)[-!#$%&'*+/0-9=?A-Z^_`a-z{|}~]+(\.[-!#$%&'*+/0-9=?A-Z^_`a-z{|}~]+)*@[A-Za-z0-9]([A-Za-z0-9-]{0,61}[A-Za-z0-9])?(\.[A-Za-z0-9]([A-Za-z0-9-]{0,61}[A-Za-z0-9])?)*$/
-export class Email extends VOString({ trim: true, maxLength: 256, pattern: EMAIL_PATTERN }) {}