+++ CHANGELOG.md
@@ -6,10 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [1.3.0](https://github.com/LucasPaganini/value-objects/releases/tag/v1.3.0) - 2020-10-03
+
 ### Added
 
 - Generate a website for the API docs
 - Use Docker for the website
+
+### Changed
+
+- #5 Errors show path to property
+- #10 Remove `lib/` for common value objects
 
 ## [1.2.1](https://github.com/LucasPaganini/value-objects/releases/tag/v1.2.1) - 2020-06-08